Split files into directories
