+++ dynamic_array/string.c
