+++ parser/parser.h
@@ -1,0 +1,11 @@
+#ifndef _PARSER_H_
+#define _PARSER_H_
+
+#include "../dynamic_array/string.h"
+#include "./lexer/lexer.h"
+#include "./ast_build/ast_parser.h"
+#include "../types.h"
+#include "../debug_macroc.h"
+#include "./ast_print/ast_print.h"
+
+#endif
